+++ docs/app.js
@@ -32,12 +32,6 @@ new Vue({
     el: '#app',
     data: {
         http: new Http(),
-        perPageOptions: [
-            {label: '3 users', value: 3},
-            {label: '5 users', value: 5},
-            {label: '10 users', value: 10},
-            {label: '12 users', value: 12},
-        ],
     },
     components: {AsyncSearch},
 })